api/requirements.txt
flask==3.1.1
python-dotenv==1.1.0
boto3==1.38.40
b2sdk==2.9.4
Werkzeug==3.1.3
flask-cors==6.0.1 